checkpoint: creating travis script
