+++ .eslintignore
@@ -3,3 +3,4 @@ fixtures/
 dist/
 .DS_Store
 node_modules/
+tmp/
