+++ .travis.yml
@@ -1,4 +1,15 @@
 sudo: required
 services:
 - docker
-script: docker-compose -f docker-compose.ci.yml run --rm generator-spike
+install:
+  - docker-compose -f docker-compose.ci.yml pull hub
+  - docker-compose -f docker-compose.ci.yml pull chrome
+  - docker build --pull --no-cache --tag spike:${BRANCH_NAME}_${BUILD_NUMBER} .
+script:
+  - docker-compose -f docker-compose.ci.yml up -d chrome
+  - docker-compose -f docker-compose.ci.yml run -d generator-spike
+  - docker-compose -f docker-compose.ci.yml run --rm generator-spike sh ./scripts/isTmpAppReady.sh && sh ./scripts/isSeleniumHubReady.sh
+  - docker-compose -f docker-compose.ci.yml run --rm generator-spike npm run test:integration
+after_script:
+  - docker-compose -f docker-compose.ci.yml down
+  - docker rmi spike:${BRANCH_NAME}_${BUILD_NUMBER}
