+++ cleanup.sh
@@ -1,0 +1,4 @@
+export BRANCH_NAME=spike
+export BUILD_NUMBER=10
+docker-compose -f docker-compose.ci.yml down
+docker rmi spike:spike_10
