+++ config/default.js
@@ -2,5 +2,6 @@ module.exports = {
   integrationTests: {
     origin: 'http://127.0.0.1:3000',
     browser: 'chrome'
-  }
+  },
+  unitTestInDir: 'tmp/unit'
 };
