+++ doit.sh
@@ -1,0 +1,7 @@
+export BRANCH_NAME=spike
+export BUILD_NUMBER=10
+docker-compose -f docker-compose.ci.yml pull hub &&\
+  docker-compose -f docker-compose.ci.yml pull chrome &&\
+  docker build --pull --no-cache --tag spike:${BRANCH_NAME}_${BUILD_NUMBER} . &&\
+  docker-compose -f docker-compose.ci.yml up -d &&\
+  docker-compose -f docker-compose.ci.yml run --rm -w"/home/spike/generator" spike npm run test:integration
