+++ scripts/build.sh
@@ -1,8 +1,6 @@
 cd "${0%/*}" && cd ..
 rm -r dist || mkdir dist
-cd src
-rsync --exclude **/dist --exclude **/node_modules --recursive --relative **/templates ../dist
-cd ..
 npm run babel -- src -d dist --ignore test.js,templates/,fixtures/
+npm run copy:templates
 cp package.json dist
 cp README.md dist
