+++ src/reducer/fixtures/tc1/src/shared/redux/reducers/something/index.js
@@ -1,0 +1,16 @@
+import { createReducer } from 'redux-act';
+import actions from 'shared/redux/actions';
+
+export default createReducer({
+  [actions.user.juggle.getType()]: (state, payload) => ({
+    ...state,
+  }),
+  [actions.authentication.login.getType()]: (state, payload) => ({
+    ...state,
+  }),
+  
+}, { 
+  token: null,
+  username: 'User',
+   
+});
